+++ engine/language-server/src/main/scala/org/enso/languageserver/requesthandler/capability/ReleaseCapabilityHandler.scala
@@ -1,29 +1,32 @@
-package org.enso.languageserver.requesthandler
+package org.enso.languageserver.requesthandler.capability
 
 import akka.actor.{Actor, ActorLogging, ActorRef, Cancellable, Props}
 import org.enso.jsonrpc.Errors.ServiceError
 import org.enso.jsonrpc._
-import org.enso.languageserver.capability.CapabilityApi.AcquireCapability
+import org.enso.languageserver.capability.CapabilityApi.{
+  CapabilityNotAcquired,
+  ReleaseCapability
+}
 import org.enso.languageserver.capability.CapabilityProtocol
 import org.enso.languageserver.capability.CapabilityProtocol.{
-  CapabilityAcquired,
-  CapabilityAcquisitionBadRequest,
-  CapabilityAcquisitionFileSystemFailure
+  CapabilityNotAcquiredResponse,
+  CapabilityReleaseBadRequest,
+  CapabilityReleased
 }
 import org.enso.languageserver.data.{CapabilityRegistration, Client}
-import org.enso.languageserver.filemanager.FileSystemFailureMapper
+import org.enso.languageserver.requesthandler.RequestTimeout
 import org.enso.languageserver.util.UnhandledLogging
 
 import scala.concurrent.duration.FiniteDuration
 
 /**
-  * A request handler for `capability/acquire` commands.
+  * A request handler for `capability/release` commands.
   *
   * @param capabilityRouter a router that dispatches capability requests
   * @param timeout a request timeout
   * @param client an object representing a client connected to the language server
   */
-class AcquireCapabilityHandler(
+class ReleaseCapabilityHandler(
   capabilityRouter: ActorRef,
   timeout: FiniteDuration,
   client: Client
@@ -31,55 +34,48 @@ class AcquireCapabilityHandler(
     with ActorLogging
     with UnhandledLogging {
 
-  import context.dispatcher
-
   override def receive: Receive = requestStage
 
+  import context.dispatcher
+
   private def requestStage: Receive = {
-    case Request(AcquireCapability, id, registration: CapabilityRegistration) =>
-      capabilityRouter ! CapabilityProtocol.AcquireCapability(
-        client,
-        registration
-      )
+    case Request(ReleaseCapability, id, params: CapabilityRegistration) =>
+      capabilityRouter ! CapabilityProtocol.ReleaseCapability(client, params)
       val cancellable =
         context.system.scheduler.scheduleOnce(timeout, self, RequestTimeout)
       context.become(responseStage(id, sender(), cancellable))
   }
-
   private def responseStage(
     id: Id,
     replyTo: ActorRef,
     cancellable: Cancellable
   ): Receive = {
     case RequestTimeout =>
-      log.error(s"Acquiring capability for ${client.id} timed out")
+      log.error(s"Releasing capability for ${client.id} timed out")
       replyTo ! ResponseError(Some(id), ServiceError)
       context.stop(self)
 
-    case CapabilityAcquired =>
-      replyTo ! ResponseResult(AcquireCapability, id, Unused)
+    case CapabilityReleased =>
+      replyTo ! ResponseResult(ReleaseCapability, id, Unused)
       cancellable.cancel()
       context.stop(self)
 
-    case CapabilityAcquisitionBadRequest =>
+    case CapabilityReleaseBadRequest =>
       replyTo ! ResponseError(Some(id), ServiceError)
       cancellable.cancel()
       context.stop(self)
 
-    case CapabilityAcquisitionFileSystemFailure(error) =>
-      replyTo ! ResponseError(
-        Some(id),
-        FileSystemFailureMapper.mapFailure(error)
-      )
+    case CapabilityNotAcquiredResponse =>
+      replyTo ! ResponseError(Some(id), CapabilityNotAcquired)
       cancellable.cancel()
       context.stop(self)
   }
 }
 
-object AcquireCapabilityHandler {
+object ReleaseCapabilityHandler {
 
   /**
-    * Creates a configuration object used to create a [[AcquireCapabilityHandler]]
+    * Creates a configuration object used to create a [[ReleaseCapabilityHandler]]
     *
     * @param capabilityRouter a router that dispatches capability requests
     * @param requestTimeout a request timeout
@@ -92,7 +88,7 @@ object AcquireCapabilityHandler {
     client: Client
   ): Props =
     Props(
-      new AcquireCapabilityHandler(capabilityRouter, requestTimeout, client)
+      new ReleaseCapabilityHandler(capabilityRouter, requestTimeout, client)
     )
 
 }
