+++ engine/language-server/src/main/scala/org/enso/languageserver/requesthandler/text/CloseFileHandler.scala
@@ -1,13 +1,14 @@
-package org.enso.languageserver.requesthandler
+package org.enso.languageserver.requesthandler.text
 
 import akka.actor.{Actor, ActorLogging, ActorRef, Cancellable, Props}
 import org.enso.jsonrpc.Errors.ServiceError
 import org.enso.jsonrpc._
 import org.enso.languageserver.data.Client
-import org.enso.languageserver.util.UnhandledLogging
+import org.enso.languageserver.requesthandler.RequestTimeout
 import org.enso.languageserver.text.TextApi.{CloseFile, FileNotOpenedError}
 import org.enso.languageserver.text.TextProtocol
 import org.enso.languageserver.text.TextProtocol.{FileClosed, FileNotOpened}
+import org.enso.languageserver.util.UnhandledLogging
 
 import scala.concurrent.duration.FiniteDuration
