+++ engine/language-server/src/main/scala/org/enso/languageserver/requesthandler/text/OpenFileHandler.scala
@@ -1,17 +1,18 @@
-package org.enso.languageserver.requesthandler
+package org.enso.languageserver.requesthandler.text
 
 import akka.actor.{Actor, ActorLogging, ActorRef, Cancellable, Props}
 import org.enso.jsonrpc.Errors.ServiceError
 import org.enso.jsonrpc.{Id, Request, ResponseError, ResponseResult}
 import org.enso.languageserver.data.Client
 import org.enso.languageserver.filemanager.FileSystemFailureMapper
-import org.enso.languageserver.util.UnhandledLogging
+import org.enso.languageserver.requesthandler.RequestTimeout
 import org.enso.languageserver.text.TextApi.OpenFile
 import org.enso.languageserver.text.TextProtocol
 import org.enso.languageserver.text.TextProtocol.{
   OpenFileResponse,
   OpenFileResult
 }
+import org.enso.languageserver.util.UnhandledLogging
 
 import scala.concurrent.duration.FiniteDuration
