+++ engine/language-server/src/main/scala/org/enso/languageserver/requesthandler/text/SaveFileHandler.scala
@@ -1,24 +1,31 @@
-package org.enso.languageserver.requesthandler
+package org.enso.languageserver.requesthandler.text
 
 import akka.actor.{Actor, ActorLogging, ActorRef, Cancellable, Props}
 import org.enso.jsonrpc.Errors.ServiceError
 import org.enso.jsonrpc._
 import org.enso.languageserver.data.Client
-import org.enso.languageserver.util.UnhandledLogging
-import org.enso.languageserver.text.TextApi._
+import org.enso.languageserver.filemanager.FileSystemFailureMapper
+import org.enso.languageserver.requesthandler.RequestTimeout
+import org.enso.languageserver.text.TextApi.{
+  FileNotOpenedError,
+  InvalidVersionError,
+  SaveFile,
+  WriteDeniedError
+}
 import org.enso.languageserver.text.TextProtocol
-import org.enso.languageserver.text.TextProtocol.{ApplyEdit => _, _}
+import org.enso.languageserver.text.TextProtocol._
+import org.enso.languageserver.util.UnhandledLogging
 
 import scala.concurrent.duration.FiniteDuration
 
 /**
-  * A request handler for `text/applyEdit` commands.
+  * A request handler for `text/save` commands.
   *
   * @param bufferRegistry a router that dispatches text editing requests
   * @param timeout a request timeout
   * @param client an object representing a client connected to the language server
   */
-class ApplyEditHandler(
+class SaveFileHandler(
   bufferRegistry: ActorRef,
   timeout: FiniteDuration,
   client: Client
@@ -31,8 +38,12 @@ class ApplyEditHandler(
   override def receive: Receive = requestStage
 
   private def requestStage: Receive = {
-    case Request(ApplyEdit, id, params: ApplyEdit.Params) =>
-      bufferRegistry ! TextProtocol.ApplyEdit(client.id, params.edit)
+    case Request(SaveFile, id, params: SaveFile.Params) =>
+      bufferRegistry ! TextProtocol.SaveFile(
+        client.id,
+        params.path,
+        params.currentVersion
+      )
       val cancellable =
         context.system.scheduler.scheduleOnce(timeout, self, RequestTimeout)
       context.become(responseStage(id, sender(), cancellable))
@@ -44,30 +55,33 @@ class ApplyEditHandler(
     cancellable: Cancellable
   ): Receive = {
     case RequestTimeout =>
-      log.error(s"Applying edit for ${client.id} timed out")
+      log.error(s"Saving file for ${client.id} timed out")
       replyTo ! ResponseError(Some(id), ServiceError)
       context.stop(self)
 
-    case ApplyEditSuccess =>
-      replyTo ! ResponseResult(ApplyEdit, id, Unused)
+    case FileSaved =>
+      replyTo ! ResponseResult(SaveFile, id, Unused)
       cancellable.cancel()
       context.stop(self)
 
-    case TextEditValidationFailed(msg) =>
-      replyTo ! ResponseError(Some(id), TextEditValidationError(msg))
-      cancellable.cancel()
-      context.stop(self)
-
-    case TextEditInvalidVersion(clientVersion, serverVersion) =>
+    case SaveFailed(fsFailure) =>
       replyTo ! ResponseError(
         Some(id),
-        InvalidVersionError(clientVersion, serverVersion)
+        FileSystemFailureMapper.mapFailure(fsFailure)
       )
       cancellable.cancel()
       context.stop(self)
 
-    case WriteDenied =>
+    case SaveDenied =>
       replyTo ! ResponseError(Some(id), WriteDeniedError)
+      cancellable.cancel()
+      context.stop(self)
+
+    case SaveFileInvalidVersion(clientVersion, serverVersion) =>
+      replyTo ! ResponseError(
+        Some(id),
+        InvalidVersionError(clientVersion, serverVersion)
+      )
       cancellable.cancel()
       context.stop(self)
 
@@ -78,10 +92,10 @@ class ApplyEditHandler(
   }
 }
 
-object ApplyEditHandler {
+object SaveFileHandler {
 
   /**
-    * Creates a configuration object used to create a [[ApplyEditHandler]]
+    * Creates a configuration object used to create a [[SaveFileHandler]].
     *
     * @param bufferRegistry a router that dispatches text editing requests
     * @param requestTimeout a request timeout
@@ -92,6 +106,6 @@ object ApplyEditHandler {
     bufferRegistry: ActorRef,
     requestTimeout: FiniteDuration,
     client: Client
-  ): Props = Props(new ApplyEditHandler(bufferRegistry, requestTimeout, client))
+  ): Props = Props(new SaveFileHandler(bufferRegistry, requestTimeout, client))
 
 }
